Complete

+++ angular/src/app/employee/employee-show/employee-show.component.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
 import { Location } from '@angular/common';
 import { EmployeeService } from '../employee.service';
 import { Employee } from "../employee";
+import {MessageService} from "../../shared/message/message.service";
 
 @Component({
   selector: 'app-employee-show',
@@ -12,15 +13,24 @@ import { Employee } from "../employee";
 export class EmployeeShowComponent implements OnInit {
 
   item?: Employee;
+  private clearMessagesOnDistroy = true;
 
   constructor(
     private route: ActivatedRoute,
     private employeeService: EmployeeService,
+    private messageService: MessageService,
     private router: Router,
   ) {}
 
   ngOnInit(): void {
+    this.clearMessagesOnDistroy = true;
     this.getItem();
+  }
+
+  ngOnDestroy(): void {
+    if (this.clearMessagesOnDistroy) {
+      this.messageService.clear();
+    }
   }
 
   getItem(): void {
@@ -30,7 +40,12 @@ export class EmployeeShowComponent implements OnInit {
   }
 
   delete(item: Employee): void {
+    if (!window.confirm("Are you sure you want to delete this item?"))
+      return;
+
+    this.messageService.clear();
     this.employeeService.deleteItem(item).subscribe();
+    this.clearMessagesOnDistroy = false;
     this.router.navigate(["/employees"])
   }
 }
