+++ angular/src/app/employee/employee-update/employee-update.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { EmployeeService } from '../employee.service';
 import { Employee } from "../employee";
+import {MessageService} from "../../shared/message/message.service";
 
 @Component({
   selector: 'app-employee-update',
@@ -11,15 +12,24 @@ import { Employee } from "../employee";
 export class EmployeeUpdateComponent implements OnInit {
 
   item?: Employee;
+  private clearMessagesOnDistroy = true;
 
   constructor(
     private route: ActivatedRoute,
     private employeeService: EmployeeService,
+    private messageService: MessageService,
     private router: Router,
   ) {}
 
   ngOnInit(): void {
+    this.clearMessagesOnDistroy = true;
     this.getItem();
+  }
+
+  ngOnDestroy(): void {
+    if (this.clearMessagesOnDistroy) {
+      this.messageService.clear();
+    }
   }
 
   getItem(): void {
@@ -29,11 +39,18 @@ export class EmployeeUpdateComponent implements OnInit {
   }
 
   delete(item: Employee): void {
+    if (!window.confirm("Are you sure you want to delete this item?"))
+      return;
+
+    this.messageService.clear();
     this.employeeService.deleteItem(item).subscribe();
+    this.clearMessagesOnDistroy = false;
+    this.router.navigate(["/employees"])
   }
 
   itemUpdated(item: Employee) {
     //this.item = item;
+    this.clearMessagesOnDistroy = false;
     this.router.navigate(["/employees"])
   }
 }
